Implementation
Begun created the main() function for how the program runs using several of the already created one. As time goes on we'll discover those that can be done without.
I've used an alternative to the json issue by creating a js waypoint file instead just so we can proceed with development and test the program. I have also created a branch Kojo-Main because there may be some severe changes