Distance fine-tuning
The app now uses the getDistance function to determing if the user is approaching a waypoint and also whether the user is within 60m of their previous location.
This should stop constant map generation until the user has moved a reasonable distance (>60m).
I have tested the program in a live environment by walking around while holding it to see if it would work and there's a single issue that stands out:
The general accuracy of the GPS device is an issue for us that I'm not sure how we can resolve as of now. I've tried setting the accurary parameter for getCurrentPosition function to true but that also generates a host of issues